requirements.txt
djangorestframework~=3.13.1
python-decouple
drf-yasg
Pillow
psycopg2-binary
django-filter
transliterate
twilio
djangorestframework-simplejwt
redis
celery
python-slugify
gunicorn
whitenoise~=6.2.0
dj-database-url
django-cors-headers
